refactor: deleteVoteByGroupId 메서드를 VoteCleanerService로 분리 (#280)
